+++ src/rule.ml
@@ -22,15 +22,19 @@ type t =
   ; env      : Env.t option
   ; build    : (unit, Action.t) Build.t
   ; targets  : Path.Build.Set.t
-  ; sandbox  : bool
   ; mode     : Dune_file.Rule.Mode.t
   ; locks    : Path.t list
   ; info     : Info.t
   ; dir      : Path.Build.t
   }
 
-let make ?(sandbox=false) ?(mode=Dune_file.Rule.Mode.Standard)
+let make
+      ?(sandbox = Sandbox_config.default)
+      ?(mode=Dune_file.Rule.Mode.Standard)
       ~context ~env ?(locks=[]) ?(info=Info.Internal) build =
+  let build =
+    Build.S.seq (Build.dep (Dep.sandbox_config sandbox)) build
+  in
   let targets = Build.targets build in
   let dir =
     match Path.Build.Set.choose targets with
@@ -64,7 +68,6 @@ let make ?(sandbox=false) ?(mode=Dune_file.Rule.Mode.Standard)
   ; env
   ; build
   ; targets
-  ; sandbox
   ; mode
   ; locks
   ; info